@@ -1,8 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.1</TargetFramework>
         <IsPackable>false</IsPackable>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
     </PropertyGroup>
 
     <ItemGroup>
@@ -13,34 +13,16 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\Moyasar\Moyasar.csproj" />
+        <ProjectReference Include="..\Moyasar\Moyasar.csproj" />
     </ItemGroup>
 
     <ItemGroup>
-      <None Remove="TestData\ObjectNotFound.json" />
-      <Content Include="Fixtures\ObjectNotFound.json">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </Content>
-      <None Remove="Fixtures\CcSourceResult.json" />
-      <Content Include="Fixtures\CcSourceResult.json">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </Content>
-      <None Remove="Fixtures\Payment.json" />
-      <Content Include="Fixtures\Payment.json">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </Content>
-      <None Remove="Fixtures\SadadSourceResult.json" />
-      <Content Include="Fixtures\SadadSourceResult.json">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </Content>
-      <None Remove="Fixtures\Invoice.json" />
-      <Content Include="Fixtures\Invoice.json">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </Content>
-      <None Remove="xunit.runner.json" />
-      <Content Include="xunit.runner.json">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </Content>
+        <Content Include="xunit.runner.json">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </Content>
+        <Content Include="Fixtures/**/*">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </Content>
     </ItemGroup>
 
 </Project>
