@@ -1,7 +1,9 @@
+using System;
+
 namespace Moyasar.Abstraction
 {
     public interface IPaymentMethod
     {
-        
+        string Type { get; }
     }
 }
