@@ -13,7 +13,8 @@ namespace Moyasar.Helpers
         static CreditCardHelper()
         {
             VisaNumberPattern = new Regex(@"^4[0-9]{12}(?:[0-9]{3})?$");
-            MasterCardNumberPattern = new Regex(@"^(?:5[1-5][0-9]{2}|222[1-9]|22[3-9][0-9]|2[3-6][0-9]{2}|27[01][0-9]|2720)[0-9]{12}$");
+            MasterCardNumberPattern =
+                new Regex(@"^(?:5[1-5][0-9]{2}|222[1-9]|22[3-9][0-9]|2[3-6][0-9]{2}|27[01][0-9]|2720)[0-9]{12}$");
         }
 
         /// <summary>
@@ -43,11 +44,21 @@ namespace Moyasar.Helpers
         /// <returns>Credit card type</returns>
         public static CreditCardType? GetCreditCardType(string number)
         {
-            if (number == null) return null;
-            
-            if (IsVisa(number)) return CreditCardType.Visa;
-            if (IsMasterCard(number)) return CreditCardType.MasterCard;
-            
+            if (number == null)
+            {
+                return null;
+            }
+
+            if (IsVisa(number))
+            {
+                return CreditCardType.Visa;
+            }
+
+            if (IsMasterCard(number))
+            {
+                return CreditCardType.MasterCard;
+            }
+
             return null;
         }
     }
