@@ -11,43 +11,22 @@ namespace Moyasar.Models
     /// <typeparam name="TModel">Resource Type</typeparam>
     public class PaginationResult<TModel> where TModel : Resource<TModel>
     {
-        private const string CurrentPageField = "current_page";
-        private const string NextPageField = "next_page";
-        private const string PreviousPageField = "prev_page";
-        private const string TotalPagesField = "total_pages";
-        private const string TotalCountField = "total_count";
-        
-        [JsonProperty(CurrentPageField)]
-        public int CurrentPage { get; set; }
-        
-        [JsonProperty(NextPageField)]
-        public int? NextPage { get; set; }
-        
-        [JsonProperty(PreviousPageField)]
-        public int? PreviousPage { get; set; }
-        
-        [JsonProperty(TotalPagesField)]
-        public int TotalPages { get; set; }
-        
-        [JsonProperty(TotalCountField)]
-        public int TotalCount { get; set; }
+        public int CurrentPage => Meta.CurrentPage;
+        public int? NextPage => Meta.NextPage;
+        public int? PreviousPage => Meta.PreviousPage;
+        public int TotalPages => Meta.TotalPages;
+        public int TotalCount => Meta.TotalCount;
 
-        [JsonIgnore]
-        internal Func<int, PaginationResult<TModel>> Paginator { get; set; }
-        
-        public PaginationResult<TModel> GetNextPage()
-        {
-            if (Paginator == null) return null;
-            return NextPage.HasValue ? Paginator(NextPage.Value) : null;
-        }
-        
-        public PaginationResult<TModel> GetPreviousPage()
-        {
-            if (Paginator == null) return null;
-            return PreviousPage.HasValue ? Paginator(PreviousPage.Value) : null;   
-        }
+        [JsonProperty("meta")]
+        public PaginationResultMeta Meta { get; private set; }
         
         [JsonIgnore]
-        public List<TModel> Items { get; set; }
+        public List<TModel> Items { get; private set; }
+
+        [JsonProperty("payments")]
+        private List<TModel> Payments { set => Items = value; }
+
+        [JsonProperty("invoices")]
+        private List<TModel> Invoices { set => Items = value; }
     }
 }
